requirements.txt
flask
bs4
pdf2image
Pillow
